+++ nncf/telemetry/__init__.py
@@ -10,3 +10,5 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 """
+
+from nncf.telemetry.wrapper import NNCFTelemetry
