+++ nncf/telemetry/events.py
@@ -1,0 +1,28 @@
+"""
+ Copyright (c) 2022 Intel Corporation
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+      http://www.apache.org/licenses/LICENSE-2.0
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+"""
+from typing import List
+
+from nncf.api.compression import CompressionAlgorithmBuilder
+from nncf.common.composite_compression import CompositeCompressionAlgorithmBuilder
+
+NNCF_TF_CATEGORY = 'nncf_tf'
+NNCF_PT_CATEGORY = 'nncf_pt'
+NNCF_ONNX_CATEGORY = 'nncf_onnx'
+NNCF_OV_CATEGORY = 'nncf_ov'
+
+def get_algo_names_from_builder(builder: CompressionAlgorithmBuilder) -> List[str]:
+    if isinstance(builder, CompositeCompressionAlgorithmBuilder):
+        algo_names = sorted([x.name for x in builder.child_builders])
+    else:
+        algo_names = [builder.name, ]
+    return algo_names
