+++ nncf/telemetry/wrapper.py
@@ -14,8 +14,6 @@ import os
 from typing import List
 
 from nncf import __version__
-from nncf.api.compression import CompressionAlgorithmBuilder
-from nncf.common.composite_compression import CompositeCompressionAlgorithmBuilder
 from nncf.common.utils.logger import logger as nncf_logger
 from nncf.definitions import NNCF_CI_ENV_VAR_NAME
 from nncf.definitions import NNCF_DEV_ENV_VAR_NAME
@@ -35,15 +33,3 @@ except ImportError:
 if os.getenv(NNCF_CI_ENV_VAR_NAME) or os.getenv(NNCF_DEV_ENV_VAR_NAME) or NNCFTelemetry is None:
     from unittest.mock import MagicMock
     NNCFTelemetry = MagicMock()
-
-NNCF_TF_CATEGORY = 'nncf_tf'
-NNCF_PT_CATEGORY = 'nncf_pt'
-NNCF_ONNX_CATEGORY = 'nncf_onnx'
-NNCF_OV_CATEGORY = 'nncf_ov'
-
-def get_algo_names_from_builder(builder: CompressionAlgorithmBuilder) -> List[str]:
-    if isinstance(builder, CompositeCompressionAlgorithmBuilder):
-        algo_names = sorted([x.name for x in builder.child_builders])
-    else:
-        algo_names = [builder.name, ]
-    return algo_names
