+++ tests/shared/helpers.py
@@ -142,7 +142,7 @@ class BaseTensorListComparator(ABC):
 
 
 def telemetry_send_event_test_driver(mocker, use_nncf_fn: Callable):
-    from nncf.telemetry_wrapper.telemetry import NNCFTelemetry
+    from nncf.telemetry import NNCFTelemetry
     telemetry_send_event_spy = mocker.spy(NNCFTelemetry, "send_event")
     use_nncf_fn()
     telemetry_send_event_spy.assert_called()
